Updated schema.sql

+++ app/src/main/resources/schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE urls (
 CREATE TABLE url_checks (
     id BIGINT GENERATED ALWAYS AS IDENTITY UNIQUE NOT NULL,
     url_id BIGINT REFERENCES urls (id) NOT NULL,
-    status_code BIGINT,
+    status_code INTEGER,
     h1 VARCHAR(255),
     title VARCHAR(255),
     description TEXT,
